LPS-171715 search-experiences-web: Separate util functions
