+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/edit_sxp_blueprint/clause_contributors_sidebar/index.js
@@ -23,7 +23,7 @@ import {
 	DESCENDING,
 	INACTIVE,
 } from '../../utils/constants';
-import {removeDuplicates} from '../../utils/utils';
+import removeDuplicates from '../../utils/functions/remove_duplicates';
 import LearnMessage from './../../shared/LearnMessage';
 import ManagementToolbar from './ManagementToolbar';
