+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/edit_sxp_blueprint/query_builder_tab/QuerySXPElements.js
@@ -19,7 +19,7 @@ import JSONSXPElement from '../../shared/JSONSXPElement';
 import SXPElement from '../../shared/sxp_element/index';
 import {SXP_ELEMENT_PREFIX} from '../../utils/constants';
 import {setStorageAddSXPElementSidebar} from '../../utils/sessionStorage';
-import {isCustomJSONSXPElement} from '../../utils/utils';
+import isCustomJSONSXPElement from '../../utils/sxp_element/is_custom_json_sxp_element';
 
 function QuerySXPElements({
 	elementInstances,
