+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/shared/PageToolbar.js
@@ -24,12 +24,12 @@ import getCN from 'classnames';
 import PropTypes from 'prop-types';
 import React, {useContext, useRef, useState} from 'react';
 
+import removeDuplicates from '../utils/functions/remove_duplicates';
 import {
 	formatLocaleWithDashes,
 	formatLocaleWithUnderscores,
 	sub,
 } from '../utils/language';
-import {removeDuplicates} from '../utils/utils';
 import ThemeContext from './ThemeContext';
 
 /**
