+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/shared/sxp_element/FieldRow.js
@@ -19,8 +19,10 @@ import {ClayTooltipProvider} from '@clayui/tooltip';
 import fuzzy from 'fuzzy';
 import React, {useContext, useEffect, useRef, useState} from 'react';
 
-import {alphabeticalSort, stringLengthSort} from '../../utils/sort';
-import {isDefined, isEmpty} from '../../utils/utils';
+import isDefined from '../../utils/functions/is_defined';
+import isEmpty from '../../utils/functions/is_empty';
+import sortAlphabetically from '../../utils/functions/sort_alphabetically';
+import sortByStringLength from '../../utils/functions/sort_by_string_length';
 import ThemeContext from '../ThemeContext';
 
 const USER_LANGUAGE_VARIABLE = '${context.language_id}';
@@ -39,10 +41,10 @@ function filterAndSortIndexFields(indexFields, field) {
 			indexField.name.toLowerCase().includes(field.toLowerCase())
 		)
 		.sort((a, b) => {
-			const sort = stringLengthSort(a.name, b.name);
+			const sort = sortByStringLength(a.name, b.name);
 
 			if (sort === 0 || field === '') {
-				return alphabeticalSort(a.name, b.name);
+				return sortAlphabetically(a.name, b.name);
 			}
 
 			return sort;
