+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/shared/sxp_element/category_selector_input/index.js
@@ -19,7 +19,8 @@ import React, {useEffect, useState} from 'react';
 
 import useDebounceCallback from '../../../hooks/useDebounceCallback';
 import {DEFAULT_HEADERS} from '../../../utils/fetch';
-import {removeDuplicates, toNumber} from '../../../utils/utils';
+import removeDuplicates from '../../../utils/functions/remove_duplicates';
+import toNumber from '../../../utils/functions/to_number';
 import CategorySelectorModal from './CategorySelectorModal';
 
 export const FETCH_URLS = {
