+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/functions/get_results_error.js
@@ -1,0 +1,35 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ */
+
+import {DEFAULT_ERROR} from '../errorMessages';
+
+/**
+ * Used for formatting a search response's error message.
+ * @param {object} error Information about the error.
+ * @returns {object}
+ */
+export default function getResultsError({
+	exceptionClass,
+	exceptionTrace,
+	msg,
+	severity,
+}) {
+	return {
+		errors: [
+			{
+				exceptionClass,
+				exceptionTrace,
+				msg: msg || DEFAULT_ERROR,
+				severity: severity || Liferay.Language.get('error'),
+			},
+		],
+	};
+}
