+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/validation.js
@@ -11,9 +11,10 @@
 
 import {ASSET_CATEGORY_ID} from './constants';
 import {ERROR_MESSAGES} from './errorMessages';
+import isDefined from './functions/is_defined';
+import isEmpty from './functions/is_empty';
 import {INPUT_TYPES} from './inputTypes';
 import {sub} from './language';
-import {isDefined, isEmpty} from './utils';
 
 export function validateBoost(configValue, type) {
 	if (configValue === null) {
