+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/view_sxp_elements/AddSXPElementModal.js
@@ -19,8 +19,8 @@ import React, {useEffect, useState} from 'react';
 
 import {CUSTOM_JSON_SXP_ELEMENT} from '../utils/data';
 import {DEFAULT_ERROR} from '../utils/errorMessages';
+import isDefined from '../utils/functions/is_defined';
 import {setInitialSuccessToast} from '../utils/toasts';
-import {isDefined} from '../utils/utils';
 
 const ADD_EVENT = 'addSXPElement';
