+++ modules/dxp/apps/search-experiences/search-experiences-web/test/js/shared/sxp_element/index.js
@@ -13,7 +13,7 @@ import {fireEvent, render} from '@testing-library/react';
 import React from 'react';
 
 import SXPElement from '../../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/shared/sxp_element/index';
-import {getUIConfigurationValues} from '../../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/utils';
+import getUIConfigurationValues from '../../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/sxp_element/get_ui_configuration_values';
 import {INDEX_FIELDS} from '../../mocks/data';
 import {QUERY_SXP_ELEMENTS} from '../../mocks/sxpElements';
