+++ modules/dxp/apps/search-experiences/search-experiences-web/test/js/utils/utils.js
@@ -9,17 +9,16 @@
  * distribution rights of the Software.
  */
 
+import isDefined from '../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/functions/is_defined';
+import isEmpty from '../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/functions/is_empty';
+import replaceStr from '../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/functions/replace_str';
 import {INPUT_TYPES} from '../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/inputTypes';
-import {
-	cleanUIConfiguration,
-	getConfigurationEntry,
+import cleanUIConfiguration from '../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/sxp_element/clean_ui_configuration';
+import getUIConfigurationValues, {
 	getDefaultValue,
-	getUIConfigurationValues,
-	isDefined,
-	isElementInactiveFromNonCompanyIndex,
-	isEmpty,
-	replaceStr,
-} from '../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/utils';
+} from '../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/sxp_element/get_ui_configuration_values';
+import isElementInactiveFromNonCompanyIndex from '../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/sxp_element/is_element_inactive_from_noncompany_index';
+import replaceTemplateVariable from '../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/sxp_element/replace_template_variable';
 import {
 	BOOST_ALL_KEYWORDS_MATCH,
 	BOOST_ASSET_TYPE,
@@ -791,10 +790,10 @@ describe('utils', () => {
 		});
 	});
 
-	describe('getConfigurationEntry', () => {
+	describe('replaceTemplateVariable', () => {
 		it('gets configurationEntry of date', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -829,7 +828,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of select', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -884,7 +883,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of itemSelector', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -920,7 +919,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of multiselect', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -953,7 +952,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of number', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -986,7 +985,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of number with suffix', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -1023,7 +1022,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of slider', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -1056,7 +1055,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of field mapping', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -1096,7 +1095,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of field mapping list', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -1160,7 +1159,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of field mapping list with undefined or blank locale', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -1217,7 +1216,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of json', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -1249,7 +1248,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of text', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -1286,7 +1285,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of configuration with multiple fields', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						elementDefinition: {
 							configuration: {
@@ -1347,7 +1346,7 @@ describe('utils', () => {
 
 		it('gets configurationEntry of custom json with no configuration', () => {
 			expect(
-				getConfigurationEntry({
+				replaceTemplateVariable({
 					sxpElement: {
 						description_i18n: {en_US: 'Editable JSON text area'},
 						elementDefinition: {
